+++ lib/helpers/formatters.ts
@@ -2,14 +2,12 @@ import { EventEmitter } from "node:events";
 
 import {
   formatterHelpers,
-  IFormatterOptions,
   JsonFormatter,
   UsageFormatter,
 } from "@cucumber/cucumber";
 import { CucumberHtmlStream } from "@cucumber/html-formatter";
 import type * as messages from "@cucumber/messages";
-import PrettyFormatter from "@cucumber/pretty-formatter";
-import chalk from "chalk";
+import { PrettyPrinter } from "@cucumber/pretty-formatter";
 
 import { assertIsString } from "./assertions";
 import { notNull } from "./type-guards";
@@ -123,67 +121,15 @@ export function createUsageFormatter(
 }
 
 export function createPrettyFormatter(
-  useColors: boolean,
-  log: (chunk: string) => void,
+  stream: NodeJS.WritableStream,
 ): EventEmitter {
   const eventBroadcaster = new EventEmitter();
 
-  const eventDataCollector = new formatterHelpers.EventDataCollector(
-    eventBroadcaster,
-  );
-
-  const colorFns: IFormatterOptions["colorFns"] = useColors
-    ? {
-        forStatus(status: messages.TestStepResultStatus) {
-          return {
-            AMBIGUOUS: chalk.red.bind(chalk),
-            FAILED: chalk.red.bind(chalk),
-            PASSED: chalk.green.bind(chalk),
-            PENDING: chalk.yellow.bind(chalk),
-            SKIPPED: chalk.cyan.bind(chalk),
-            UNDEFINED: chalk.yellow.bind(chalk),
-            UNKNOWN: chalk.yellow.bind(chalk),
-          }[status];
-        },
-        location: chalk.gray.bind(chalk),
-        tag: chalk.cyan.bind(chalk),
-        diffAdded: chalk.green.bind(chalk),
-        diffRemoved: chalk.red.bind(chalk),
-        errorMessage: chalk.red.bind(chalk),
-        errorStack: chalk.grey.bind(chalk),
-      }
-    : {
-        forStatus() {
-          return (x) => x;
-        },
-        location: (x) => x,
-        tag: (x) => x,
-        diffAdded: (x) => x,
-        diffRemoved: (x) => x,
-        errorMessage: (x) => x,
-        errorStack: (x) => x,
-      };
-
-  new PrettyFormatter({
-    eventBroadcaster,
-    eventDataCollector,
-    log(chunk) {
-      assertIsString(
-        chunk,
-        "Expected a JSON output of string, but got " + typeof chunk,
-      );
-      log(chunk);
-    },
-    supportCodeLibrary: null as any,
-    colorFns,
-    cwd: null as any,
-    parsedArgvOptions: {
-      colorsEnabled: useColors,
-    },
-    snippetBuilder: null as any,
-    stream: null as any,
-    cleanup: null as any,
+  const printer = new PrettyPrinter({
+    stream,
   });
+
+  eventBroadcaster.on("envelope", (envelope) => printer.update(envelope));
 
   return eventBroadcaster;
 }